Base: Use Darwin os_sync primitives for spinlocks
The fallback policy we use for POSIX platforms involves sleeping,
potentially for up to 4 milliseconds.  This can result in high
tail-latency should the lock become available after the thread has begun
to sleep.

Instead, let's use os_sync primitives on Darwin, not unlike how the
futex primitives are used on Linux. This allows the waiter to
immediately become eligible to acquire when the lock becomes available.